+++ R/data-static.R
@@ -34,5 +34,5 @@ new_static_block <- function(data, ...) {
 
 #' @export
 as.list.static_block <- function(x, ...) {
-  stop("Cannot serialize upload blocks.")
+  stop("Cannot serialize static blocks.")
 }