+++ tests/testthat/test-transform-fixed.R
@@ -1,0 +1,24 @@
+test_that("fixed block constructor", {
+
+  blk <- new_fixed_block(quote(head(data)))
+
+  expect_s3_class(blk, "fixed_block")
+
+  expect_error(new_fixed_block())
+  expect_error(new_fixed_block(123), class = "invalid_block_input")
+
+  testServer(
+    get_s3_method("block_server", blk),
+    {
+      session$flushReact()
+      expect_identical(
+        session$returned$result(),
+        head(datasets::mtcars)
+      )
+    },
+    args = list(
+      x = blk,
+      data = list(data = function() datasets::mtcars)
+    )
+  )
+})